refactored the connection handler class
- refactored the connection handler class and moved it as an inner class under web server to be able to access server class variables without passing a reference to the web server
