+++ src/webserver/WebServer.java
@@ -6,7 +6,12 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.HashMap;
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.BufferedReader;
+import java.io.PrintWriter;
 
+import dubjhandlers.ProblemHandler;
+import dubjhandlers.LeaderboardHandler;
 public class WebServer {
   /** The port that this WebServer is hosted on. **/
   private int port;
@@ -25,6 +30,8 @@ public class WebServer {
 
   public static void main(String[] args) {
     WebServer server = new WebServer(5000);
+    server.route("/problems", new ProblemHandler());
+    server.route("/leaderboard", new LeaderboardHandler());
     server.run();
   }
 
@@ -88,5 +95,52 @@ public class WebServer {
    */
   public void route(String route, RouteTarget target) {
     this.routes.put(route, target);
+  }
+
+  /**
+   * A connection handler designed to handle a connection from a specified client.
+   * <p>
+   * This class is designed to work alongside a webserver, and will manage all io
+   * with the connected client (expected to be a browser of some sort). More
+   * specifically, it will read in the client's HTTP request and return an HTTP
+   * response to the client.
+   * <p>
+   * Created <b>2021-01-08</b>.
+   * 
+   * @since 0.0.1
+   * @version 0.0.1
+   * @author Joseph Wang
+   */
+  private class ConnectionHandler implements Runnable {
+    /** The client to handle. */
+    private Socket client;
+    /** The input stream from the client. */
+    private BufferedReader input;
+    /** The output stream for the client. */
+    private PrintWriter output;
+
+    /**
+     * Constructs a new ConnectionHandler to handle a specific client.
+     * 
+     * @param client the client to handle.
+     */
+    public ConnectionHandler(Socket client) {
+      this.client = client;
+
+      try {
+        this.input = new BufferedReader(new InputStreamReader(this.client.getInputStream()));
+        this.output = new PrintWriter(this.client.getOutputStream());
+      } catch (IOException e) {
+        e.printStackTrace();
+      }
+    }
+
+    /**
+     * {@inheritDoc}
+     * <p>
+     * Starts this ConnectionHandler and begins handling a specific client.
+     */
+    public void run() {
+    }
   }
 }
